requirements.txt
altgraph==0.17
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
macholib==1.14
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==1.0.2
pyOpenSSL==20.0.1
PyYAML==5.4.1
queuelib==1.5.0
redis==3.5.3
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.3.0
urllib3==1.26.4
w3lib==1.22.0
Wand==0.6.5
xlrd==1.2.0
xlwt==1.3.0